backend/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.0.4
isort==5.10.1
#psycopg2-binary==2.8.6
Pillow==9.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
requests==2.26.0
pytest-pythonpath==0.7.3
pytz==2020.1
sqlparse==0.3.1
